WFLY-13965 PredicateCapabilityReference should be a special case of a ResourceCapabilityReference - not the other way around.
