+++ clustering/common/src/main/java/org/jboss/as/clustering/controller/ResourceCapabilityReference.java
@@ -22,8 +22,13 @@
 
 package org.jboss.as.clustering.controller;
 
-import org.jboss.as.clustering.function.Predicates;
+import java.util.function.Function;
+
 import org.jboss.as.controller.CapabilityReferenceRecorder;
+import org.jboss.as.controller.OperationContext;
+import org.jboss.as.controller.PathAddress;
+import org.jboss.as.controller.capability.RuntimeCapability;
+import org.jboss.as.controller.registry.Resource;
 import org.wildfly.clustering.service.BinaryRequirement;
 import org.wildfly.clustering.service.Requirement;
 import org.wildfly.clustering.service.UnaryRequirement;
@@ -32,7 +37,8 @@ import org.wildfly.clustering.service.UnaryRequirement;
  * {@link CapabilityReferenceRecorder} for resource-level capability references.
  * @author Paul Ferraro
  */
-public class ResourceCapabilityReference extends PredicateCapabilityReference {
+public class ResourceCapabilityReference extends AbstractCapabilityReference {
+    private final Function<PathAddress, String[]> requirementNameResolver;
 
     /**
      * Creates a new reference between the specified capability and the specified requirement
@@ -40,7 +46,7 @@ public class ResourceCapabilityReference extends PredicateCapabilityReference {
      * @param requirement the requirement of the specified capability
      */
     public ResourceCapabilityReference(Capability capability, Requirement requirement) {
-        super(capability, requirement, Predicates.always());
+        this(capability, requirement, SimpleCapabilityNameResolver.EMPTY);
     }
 
     /**
@@ -50,7 +56,7 @@ public class ResourceCapabilityReference extends PredicateCapabilityReference {
      * @param requirementNameResolver function for resolving the dynamic elements of the requirement name
      */
     public ResourceCapabilityReference(Capability capability, UnaryRequirement requirement, UnaryCapabilityNameResolver requirementNameResolver) {
-        super(capability, requirement, requirementNameResolver, Predicates.always());
+        this(capability, (Requirement) requirement, requirementNameResolver);
     }
 
     /**
@@ -60,6 +66,38 @@ public class ResourceCapabilityReference extends PredicateCapabilityReference {
      * @param requirementNameResolver function for resolving the dynamic elements of the requirement name
      */
     public ResourceCapabilityReference(Capability capability, BinaryRequirement requirement, BinaryCapabilityNameResolver requirementNameResolver) {
-        super(capability, requirement, requirementNameResolver, Predicates.always());
+        this(capability, (Requirement) requirement, requirementNameResolver);
+    }
+
+    private ResourceCapabilityReference(Capability capability, Requirement requirement, Function<PathAddress, String[]> requirementNameResolver) {
+        super(capability, requirement);
+        this.requirementNameResolver = requirementNameResolver;
+    }
+
+    @Override
+    public void addCapabilityRequirements(OperationContext context, Resource resource, String attributeName, String... values) {
+        context.registerAdditionalCapabilityRequirement(this.getRequirementName(context), this.getDependentName(context), attributeName);
+    }
+
+    @Override
+    public void removeCapabilityRequirements(OperationContext context, Resource resource, String attributeName, String... values) {
+        context.deregisterCapabilityRequirement(this.getRequirementName(context), this.getDependentName(context));
+    }
+
+    private String getRequirementName(OperationContext context) {
+        String[] parts = this.requirementNameResolver.apply(context.getCurrentAddress());
+        return (parts.length > 0) ? RuntimeCapability.buildDynamicCapabilityName(this.getBaseRequirementName(), parts) : this.getBaseRequirementName();
+    }
+
+    @Override
+    public String[] getRequirementPatternSegments(String name, PathAddress address) {
+        String[] segments = this.requirementNameResolver.apply(address);
+        for (int i = 0; i < segments.length; ++i) {
+            String segment = segments[i];
+            if (segment.charAt(0) == '$') {
+                segments[i] = segment.substring(1);
+            }
+        }
+        return segments;
     }
 }
